# apps/api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "DatasetCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "version"
        ],
        "title": "DatasetCreate",
        "type": "object"
      },
      "DatasetRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "name",
          "version",
          "created_at"
        ],
        "title": "DatasetRead",
        "type": "object"
      },
      "EvalCaseCreate": {
        "properties": {
          "expected": {
            "additionalProperties": true,
            "title": "Expected",
            "type": "object"
          },
          "input_payload": {
            "additionalProperties": true,
            "title": "Input Payload",
            "type": "object"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "input_payload"
        ],
        "title": "EvalCaseCreate",
        "type": "object"
      },
      "EvalCaseRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "expected": {
            "additionalProperties": true,
            "title": "Expected",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_payload": {
            "additionalProperties": true,
            "title": "Input Payload",
            "type": "object"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "input_payload",
          "expected",
          "tags",
          "created_at"
        ],
        "title": "EvalCaseRead",
        "type": "object"
      },
      "EvalCaseResultRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "eval_case_id": {
            "format": "uuid",
            "title": "Eval Case Id",
            "type": "string"
          },
          "eval_run_id": {
            "format": "uuid",
            "title": "Eval Run Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scores": {
            "additionalProperties": true,
            "title": "Scores",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/EvalCaseStatus"
          },
          "trace_grade": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Grade"
          }
        },
        "required": [
          "id",
          "eval_run_id",
          "eval_case_id",
          "status",
          "scores",
          "created_at"
        ],
        "title": "EvalCaseResultRead",
        "type": "object"
      },
      "EvalCaseStatus": {
        "enum": [
          "pending",
          "succeeded",
          "failed"
        ],
        "title": "EvalCaseStatus",
        "type": "string"
      },
      "EvalRunCreate": {
        "properties": {
          "baseline_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Version Id"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "workflow_version_id": {
            "format": "uuid",
            "title": "Workflow Version Id",
            "type": "string"
          }
        },
        "required": [
          "dataset_id",
          "workflow_version_id"
        ],
        "title": "EvalRunCreate",
        "type": "object"
      },
      "EvalRunRead": {
        "properties": {
          "baseline_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Version Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/EvalRunStatus"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "workflow_version_id": {
            "format": "uuid",
            "title": "Workflow Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "workflow_version_id",
          "status",
          "summary",
          "created_at"
        ],
        "title": "EvalRunRead",
        "type": "object"
      },
      "EvalRunStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "EvalRunStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Agent Reliability Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_projects__project_id__datasets_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetRead"
                  },
                  "title": "Response List Datasets Api V1 Projects  Project Id  Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "evals"
        ]
      },
      "post": {
        "operationId": "create_dataset_api_v1_projects__project_id__datasets_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dataset",
        "tags": [
          "evals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/datasets/{dataset_id}/cases": {
      "get": {
        "operationId": "list_eval_cases_api_v1_projects__project_id__datasets__dataset_id__cases_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvalCaseRead"
                  },
                  "title": "Response List Eval Cases Api V1 Projects  Project Id  Datasets  Dataset Id  Cases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Eval Cases",
        "tags": [
          "evals"
        ]
      },
      "post": {
        "operationId": "create_eval_case_api_v1_projects__project_id__datasets__dataset_id__cases_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalCaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalCaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Eval Case",
        "tags": [
          "evals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/eval-runs": {
      "get": {
        "operationId": "list_eval_runs_api_v1_projects__project_id__eval_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvalRunRead"
                  },
                  "title": "Response List Eval Runs Api V1 Projects  Project Id  Eval Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Eval Runs",
        "tags": [
          "evals"
        ]
      },
      "post": {
        "operationId": "create_eval_run_api_v1_projects__project_id__eval_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Eval Run",
        "tags": [
          "evals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/eval-runs/{eval_run_id}": {
      "get": {
        "operationId": "get_eval_run_api_v1_projects__project_id__eval_runs__eval_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "eval_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Eval Run",
        "tags": [
          "evals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/eval-runs/{eval_run_id}/enqueue": {
      "post": {
        "operationId": "enqueue_eval_api_v1_projects__project_id__eval_runs__eval_run_id__enqueue_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "eval_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue Eval",
        "tags": [
          "evals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/eval-runs/{eval_run_id}/execute": {
      "post": {
        "operationId": "execute_eval_api_v1_projects__project_id__eval_runs__eval_run_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "eval_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Eval",
        "tags": [
          "evals"
        ]
      }
    },
    "/api/v1/projects/{project_id}/eval-runs/{eval_run_id}/results": {
      "get": {
        "operationId": "list_eval_results_api_v1_projects__project_id__eval_runs__eval_run_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "eval_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvalCaseResultRead"
                  },
                  "title": "Response List Eval Results Api V1 Projects  Project Id  Eval Runs  Eval Run Id  Results Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Eval Results",
        "tags": [
          "evals"
        ]
      }
    },
...
  }
}